# openapi.yaml
openapi: 3.1.0
info:
  title: Social Neuron REST API
  version: 1.5.2
  description: |
    REST API for Social Neuron — AI-powered social media content creation, scheduling, and analytics.

    ## Authentication
    All endpoints require an API key via Bearer token:
    ```
    Authorization: Bearer snk_live_...
    ```
    Get your API key at [socialneuron.com/settings/developer](https://socialneuron.com/settings/developer).

    ## Scopes
    API keys have scopes that control access:
    - `mcp:full` — All permissions
    - `mcp:read` — Read analytics, brand, credits
    - `mcp:write` — Generate content, save plans
    - `mcp:distribute` — Schedule posts
    - `mcp:analytics` — Refresh analytics
    - `mcp:comments` — Manage comments
    - `mcp:autopilot` — Automated scheduling (Pro+)

    ## Async Operations
    Video/image generation returns `202 Accepted` with a `Location` header pointing to the job status endpoint.
    Poll `GET /v1/jobs/{id}` with the `Retry-After` interval until status is `completed` or `failed`.

    ## Rate Limits
    - Read operations: 60 requests/minute
    - Write/posting operations: 30 requests/minute
    - IP-level: 60 requests/minute (pre-auth)

    Rate-limited responses include `Retry-After` header.
  contact:
    name: Social Neuron
    url: https://socialneuron.com
    email: developers@socialneuron.com
  license:
    name: MIT
servers:
  - url: https://mcp.socialneuron.com/v1
    description: Production
  - url: http://localhost:8080/v1
...
security:
  - apiKey: []
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: "API key with snk_live_ prefix"
  schemas:
...
    Platform:
      type: string
      enum: [youtube, tiktok, instagram, twitter, linkedin, facebook, threads, bluesky]
...
paths:
...
  /analytics:
    get:
      summary: Fetch post analytics
      tags: [Analytics]
      parameters:
        - name: platform
          in: query
          schema:
            $ref: "#/components/schemas/Platform"
        - name: days
          in: query
          schema:
            type: integer
            default: 30
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Analytics data
  /analytics/refresh:
    post:
      summary: Trigger analytics refresh
      tags: [Analytics]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                platform:
                  $ref: "#/components/schemas/Platform"
      responses:
        "202":
          description: Refresh triggered
  /analytics/youtube:
    get:
      summary: YouTube channel analytics
      tags: [Analytics]
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 28
        - name: metrics
          in: query
          schema:
            type: string
            description: Comma-separated metric names
      responses:
        "200":
          description: YouTube analytics
  /analytics/insights:
    get:
      summary: AI performance insights
      tags: [Analytics]
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Performance insights
  /analytics/posting-times:
    get:
      summary: Best posting times
      tags: [Analytics]
      parameters:
        - name: platform
          in: query
          schema:
            $ref: "#/components/schemas/Platform"
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Recommended posting times

  # ── Brand ────────────────────────────────────────────────────────
  /brand:
    get:
      summary: Get brand profile
      tags: [Brand]
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Brand profile
    put:
      summary: Save/update brand profile
      tags: [Brand]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [brand_context]
              properties:
                brand_context:
                  type: object
                change_summary:
                  type: string
                changed_paths:
                  type: array
                  items:
                    type: string
                source_url:
                  type: string
                  format: uri
                extraction_method:
                  type: string
                overall_confidence:
                  type: number
                project_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Brand profile saved
  /brand/extract:
    post:
      summary: Extract brand from URL
      tags: [Brand]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
      responses:
        "200":
          description: Extracted brand profile

  # ── Plans ────────────────────────────────────────────────────────
...
  /comments:
    get:
      summary: List comments
      tags: [Comments]
      parameters:
        - name: platform
          in: query
          schema:
            $ref: "#/components/schemas/Platform"
        - name: video_id
          in: query
          schema:
            type: string
        - name: post_id
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [time, relevance]
            default: time
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of comments
    post:
      summary: Post a comment
      tags: [Comments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                video_id:
                  type: string
                post_id:
                  type: string
                text:
                  type: string
                platform:
                  $ref: "#/components/schemas/Platform"
      responses:
        "201":
          description: Comment posted
  /comments/{id}/reply:
    post:
      summary: Reply to a comment
      tags: [Comments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
      responses:
        "201":
          description: Reply posted
  /comments/{id}/moderate:
    post:
      summary: Moderate a comment
      tags: [Comments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [approve, hide, flag]
      responses:
        "200":
          description: Comment moderated
  /comments/{id}:
    delete:
      summary: Delete a comment
      tags: [Comments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Comment deleted

  # ── Tools ────────────────────────────────────────────────────────
...